Python3 Pickle Friendly
Python2 pickle formats write as bytestrings, but Python3 reads them as strings. To read a Python2 pickle in Python3, we must first read in the file as 'rb'. such as 

`pkl.load(filename, 'rb')`

and store them as 

`pkl.dump(filename, 'wb')`

Afterwards, we set the pickle package what the encoding as 'latin1' for it to read/write between Python2 and Python3. I got all of this information from

http://stackoverflow.com/questions/28218466/unpickling-a-python-2-object-with-python-3